+++ server/src/main/java/org/elasticsearch/common/time/JavaDateMathParser.java
@@ -17,13 +17,24 @@
  * under the License.
  */
 
-package org.elasticsearch.common.joda;
+package org.elasticsearch.common.time;
 
 import org.elasticsearch.ElasticsearchParseException;
-import org.joda.time.DateTimeZone;
-import org.joda.time.MutableDateTime;
-import org.joda.time.format.DateTimeFormatter;
 
+import java.time.DateTimeException;
+import java.time.DayOfWeek;
+import java.time.Instant;
+import java.time.LocalTime;
+import java.time.ZoneId;
+import java.time.ZoneOffset;
+import java.time.ZonedDateTime;
+import java.time.temporal.ChronoField;
+import java.time.temporal.TemporalAccessor;
+import java.time.temporal.TemporalAdjusters;
+import java.time.temporal.TemporalField;
+import java.time.temporal.TemporalQueries;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Objects;
 import java.util.function.LongSupplier;
 
@@ -34,23 +45,30 @@ import java.util.function.LongSupplier;
  * is appended to a datetime with the following syntax:
  * <code>||[+-/](\d+)?[yMwdhHms]</code>.
  */
-public class DateMathParser {
+public class JavaDateMathParser implements DateMathParser {
 
-    private final FormatDateTimeFormatter dateTimeFormatter;
-
-    public DateMathParser(FormatDateTimeFormatter dateTimeFormatter) {
-        Objects.requireNonNull(dateTimeFormatter);
-        this.dateTimeFormatter = dateTimeFormatter;
+    // base fields which should be used for default parsing, when we round up
+    private static final Map<TemporalField, Long> ROUND_UP_BASE_FIELDS = new HashMap<>(6);
+    {
+        ROUND_UP_BASE_FIELDS.put(ChronoField.MONTH_OF_YEAR, 1L);
+        ROUND_UP_BASE_FIELDS.put(ChronoField.DAY_OF_MONTH, 1L);
+        ROUND_UP_BASE_FIELDS.put(ChronoField.HOUR_OF_DAY, 23L);
+        ROUND_UP_BASE_FIELDS.put(ChronoField.MINUTE_OF_HOUR, 59L);
+        ROUND_UP_BASE_FIELDS.put(ChronoField.SECOND_OF_MINUTE, 59L);
+        ROUND_UP_BASE_FIELDS.put(ChronoField.MILLI_OF_SECOND, 999L);
     }
 
-    public long parse(String text, LongSupplier now) {
-        return parse(text, now, false, null);
+    private final DateFormatter formatter;
+    private final DateFormatter roundUpFormatter;
+
+    public JavaDateMathParser(DateFormatter formatter) {
+        Objects.requireNonNull(formatter);
+        this.formatter = formatter;
+        this.roundUpFormatter = formatter.parseDefaulting(ROUND_UP_BASE_FIELDS);
     }
 
-    // Note: we take a callable here for the timestamp in order to be able to figure out
-    // if it has been used. For instance, the request cache does not cache requests that make
-    // use of `now`.
-    public long parse(String text, LongSupplier now, boolean roundUp, DateTimeZone timeZone) {
+    @Override
+    public long parse(String text, LongSupplier now, boolean roundUp, ZoneId timeZone) {
         long time;
         String mathString;
         if (text.startsWith("now")) {
@@ -72,11 +90,12 @@ public class DateMathParser {
         return parseMath(mathString, time, roundUp, timeZone);
     }
 
-    private long parseMath(String mathString, long time, boolean roundUp, DateTimeZone timeZone) throws ElasticsearchParseException {
+    private long parseMath(final String mathString, final long time, final boolean roundUp,
+                           ZoneId timeZone) throws ElasticsearchParseException {
         if (timeZone == null) {
-            timeZone = DateTimeZone.UTC;
+            timeZone = ZoneOffset.UTC;
         }
-        MutableDateTime dateTime = new MutableDateTime(time, timeZone);
+        ZonedDateTime dateTime = ZonedDateTime.ofInstant(Instant.ofEpochMilli(time), timeZone);
         for (int i = 0; i < mathString.length(); ) {
             char c = mathString.charAt(i++);
             final boolean round;
@@ -118,100 +137,104 @@ public class DateMathParser {
                 }
             }
             char unit = mathString.charAt(i++);
-            MutableDateTime.Property propertyToRound = null;
             switch (unit) {
                 case 'y':
                     if (round) {
-                        propertyToRound = dateTime.yearOfCentury();
+                        dateTime = dateTime.withDayOfYear(1).with(LocalTime.MIN);
                     } else {
-                        dateTime.addYears(sign * num);
+                        dateTime = dateTime.plusYears(sign * num);
+                    }
+                    if (roundUp) {
+                        dateTime = dateTime.plusYears(1);
                     }
                     break;
                 case 'M':
                     if (round) {
-                        propertyToRound = dateTime.monthOfYear();
+                        dateTime = dateTime.withDayOfMonth(1).with(LocalTime.MIN);
                     } else {
-                        dateTime.addMonths(sign * num);
+                        dateTime = dateTime.plusMonths(sign * num);
+                    }
+                    if (roundUp) {
+                        dateTime = dateTime.plusMonths(1);
                     }
                     break;
                 case 'w':
                     if (round) {
-                        propertyToRound = dateTime.weekOfWeekyear();
+                        dateTime = dateTime.with(TemporalAdjusters.previousOrSame(DayOfWeek.MONDAY)).with(LocalTime.MIN);
                     } else {
-                        dateTime.addWeeks(sign * num);
+                        dateTime = dateTime.plusWeeks(sign * num);
+                    }
+                    if (roundUp) {
+                        dateTime = dateTime.plusWeeks(1);
                     }
                     break;
                 case 'd':
                     if (round) {
-                        propertyToRound = dateTime.dayOfMonth();
+                        dateTime = dateTime.with(LocalTime.MIN);
                     } else {
-                        dateTime.addDays(sign * num);
+                        dateTime = dateTime.plusDays(sign * num);
+                    }
+                    if (roundUp) {
+                        dateTime = dateTime.plusDays(1);
                     }
                     break;
                 case 'h':
                 case 'H':
                     if (round) {
-                        propertyToRound = dateTime.hourOfDay();
+                        dateTime = dateTime.withMinute(0).withSecond(0).withNano(0);
                     } else {
-                        dateTime.addHours(sign * num);
+                        dateTime = dateTime.plusHours(sign * num);
+                    }
+                    if (roundUp) {
+                        dateTime = dateTime.plusHours(1);
                     }
                     break;
                 case 'm':
                     if (round) {
-                        propertyToRound = dateTime.minuteOfHour();
+                        dateTime = dateTime.withSecond(0).withNano(0);
                     } else {
-                        dateTime.addMinutes(sign * num);
+                        dateTime = dateTime.plusMinutes(sign * num);
+                    }
+                    if (roundUp) {
+                        dateTime = dateTime.plusMinutes(1);
                     }
                     break;
                 case 's':
                     if (round) {
-                        propertyToRound = dateTime.secondOfMinute();
+                        dateTime = dateTime.withNano(0);
                     } else {
-                        dateTime.addSeconds(sign * num);
+                        dateTime = dateTime.plusSeconds(sign * num);
+                    }
+                    if (roundUp) {
+                        dateTime = dateTime.plusSeconds(1);
                     }
                     break;
                 default:
                     throw new ElasticsearchParseException("unit [{}] not supported for date math [{}]", unit, mathString);
             }
-            if (propertyToRound != null) {
-                if (roundUp) {
-                    // we want to go up to the next whole value, even if we are already on a rounded value
-                    propertyToRound.add(1);
-                    propertyToRound.roundFloor();
-                    dateTime.addMillis(-1); // subtract 1 millisecond to get the largest inclusive value
-                } else {
-                    propertyToRound.roundFloor();
-                }
+            if (roundUp) {
+                dateTime = dateTime.minus(1, ChronoField.MILLI_OF_SECOND.getBaseUnit());
             }
         }
-        return dateTime.getMillis();
+        return dateTime.toInstant().toEpochMilli();
     }
 
-    private long parseDateTime(String value, DateTimeZone timeZone, boolean roundUpIfNoTime) {
-        DateTimeFormatter parser = dateTimeFormatter.parser();
-        if (timeZone != null) {
-            parser = parser.withZone(timeZone);
-        }
+    private long parseDateTime(String value, ZoneId timeZone, boolean roundUpIfNoTime) {
+        DateFormatter formatter = roundUpIfNoTime ? this.roundUpFormatter : this.formatter;
         try {
-            MutableDateTime date;
-            // We use 01/01/1970 as a base date so that things keep working with date
-            // fields that are filled with times without dates
-            if (roundUpIfNoTime) {
-                date = new MutableDateTime(1970, 1, 1, 23, 59, 59, 999, DateTimeZone.UTC);
+            if (timeZone == null) {
+                return DateFormatters.toZonedDateTime(formatter.parse(value)).toInstant().toEpochMilli();
             } else {
-                date = new MutableDateTime(1970, 1, 1, 0, 0, 0, 0, DateTimeZone.UTC);
+                TemporalAccessor accessor = formatter.parse(value);
+                ZoneId zoneId = TemporalQueries.zone().queryFrom(accessor);
+                if (zoneId != null) {
+                    timeZone = zoneId;
+                }
+
+                return DateFormatters.toZonedDateTime(accessor).withZoneSameLocal(timeZone).toInstant().toEpochMilli();
             }
-            final int end = parser.parseInto(date, value, 0);
-            if (end < 0) {
-                int position = ~end;
-                throw new IllegalArgumentException("Parse failure at index [" + position + "] of [" + value + "]");
-            } else if (end != value.length()) {
-                throw new IllegalArgumentException("Unrecognized chars at the end of [" + value + "]: [" + value.substring(end) + "]");
-            }
-            return date.getMillis();
-        } catch (IllegalArgumentException e) {
-            throw new ElasticsearchParseException("failed to parse date field [{}] with format [{}]", e, value, dateTimeFormatter.format());
+        } catch (IllegalArgumentException | DateTimeException e) {
+            throw new ElasticsearchParseException("failed to parse date field [{}]: [{}]", e, value, e.getMessage());
         }
     }
-
 }
