+++ server/src/main/java/org/elasticsearch/index/mapper/DateFieldMapper.java
@@ -36,10 +36,11 @@ import org.apache.lucene.util.BytesRef;
 import org.elasticsearch.common.Explicit;
 import org.elasticsearch.common.Nullable;
 import org.elasticsearch.common.geo.ShapeRelation;
-import org.elasticsearch.common.joda.DateMathParser;
 import org.elasticsearch.common.joda.FormatDateTimeFormatter;
 import org.elasticsearch.common.joda.Joda;
+import org.elasticsearch.common.joda.JodaDateMathParser;
 import org.elasticsearch.common.settings.Settings;
+import org.elasticsearch.common.time.DateFormatters;
 import org.elasticsearch.common.util.LocaleUtils;
 import org.elasticsearch.common.xcontent.XContentBuilder;
 import org.elasticsearch.common.xcontent.support.XContentMapValues;
@@ -174,7 +175,7 @@ public class DateFieldMapper extends FieldMapper {
 
     public static final class DateFieldType extends MappedFieldType {
         protected FormatDateTimeFormatter dateTimeFormatter;
-        protected DateMathParser dateMathParser;
+        protected JodaDateMathParser dateMathParser;
 
         DateFieldType() {
             super();
@@ -231,10 +232,10 @@ public class DateFieldMapper extends FieldMapper {
         public void setDateTimeFormatter(FormatDateTimeFormatter dateTimeFormatter) {
             checkIfFrozen();
             this.dateTimeFormatter = dateTimeFormatter;
-            this.dateMathParser = new DateMathParser(dateTimeFormatter);
+            this.dateMathParser = new JodaDateMathParser(dateTimeFormatter);
         }
 
-        protected DateMathParser dateMathParser() {
+        protected JodaDateMathParser dateMathParser() {
             return dateMathParser;
         }
 
@@ -262,13 +263,13 @@ public class DateFieldMapper extends FieldMapper {
 
         @Override
         public Query rangeQuery(Object lowerTerm, Object upperTerm, boolean includeLower, boolean includeUpper, ShapeRelation relation,
-                @Nullable DateTimeZone timeZone, @Nullable DateMathParser forcedDateParser, QueryShardContext context) {
+                                @Nullable DateTimeZone timeZone, @Nullable JodaDateMathParser forcedDateParser, QueryShardContext context) {
             failIfNotIndexed();
             if (relation == ShapeRelation.DISJOINT) {
                 throw new IllegalArgumentException("Field [" + name() + "] of type [" + typeName() +
                         "] does not support DISJOINT ranges");
             }
-            DateMathParser parser = forcedDateParser == null
+            JodaDateMathParser parser = forcedDateParser == null
                     ? dateMathParser
                     : forcedDateParser;
             long l, u;
@@ -296,9 +297,9 @@ public class DateFieldMapper extends FieldMapper {
             return query;
         }
 
-        public long parseToMilliseconds(Object value, boolean roundUp,
-                @Nullable DateTimeZone zone, @Nullable DateMathParser forcedDateParser, QueryRewriteContext context) {
-            DateMathParser dateParser = dateMathParser();
+        public long parseToMilliseconds(Object value, boolean roundUp, @Nullable DateTimeZone zone,
+                                        @Nullable JodaDateMathParser forcedDateParser, QueryRewriteContext context) {
+            JodaDateMathParser dateParser = dateMathParser();
             if (forcedDateParser != null) {
                 dateParser = forcedDateParser;
             }
@@ -309,13 +310,13 @@ public class DateFieldMapper extends FieldMapper {
             } else {
                 strValue = value.toString();
             }
-            return dateParser.parse(strValue, context::nowInMillis, roundUp, zone);
+            return dateParser.parse(strValue, context::nowInMillis, roundUp, DateFormatters.timeZoneToZoneId(zone));
         }
 
         @Override
-        public Relation isFieldWithinQuery(IndexReader reader,
-                Object from, Object to, boolean includeLower, boolean includeUpper,
-                DateTimeZone timeZone, DateMathParser dateParser, QueryRewriteContext context) throws IOException {
+        public Relation isFieldWithinQuery(IndexReader reader, Object from, Object to, boolean includeLower, boolean includeUpper,
+                                           DateTimeZone timeZone, JodaDateMathParser dateParser,
+                                           QueryRewriteContext context) throws IOException {
             if (dateParser == null) {
                 dateParser = this.dateMathParser;
             }
